+++ examples/example-protocol/src/assets/ts_runtime_test/expected_index.ts
@@ -20,7 +20,8 @@ export type Imports = {
     importFpStruct: (arg: types.FpPropertyRenaming) => types.FpPropertyRenaming;
     importFpUntagged: (arg: types.FpUntagged) => types.FpUntagged;
     importGenerics: (arg: types.StructWithGenerics<number>) => types.StructWithGenerics<number>;
-    importGetBytes: () => types.Result<ArrayBuffer, string>;
+    importGetBytes: () => types.Result<Uint8Array, string>;
+    importGetSerdeBytes: () => types.Result<ArrayBuffer, string>;
     importMultiplePrimitives: (arg1: number, arg2: string) => bigint;
     importPrimitiveBool: (arg: boolean) => boolean;
     importPrimitiveF32: (arg: number) => number;
@@ -57,7 +58,8 @@ export type Exports = {
     exportFpStruct?: (arg: types.FpPropertyRenaming) => types.FpPropertyRenaming;
     exportFpUntagged?: (arg: types.FpUntagged) => types.FpUntagged;
     exportGenerics?: (arg: types.StructWithGenerics<number>) => types.StructWithGenerics<number>;
-    exportGetBytes?: () => types.Result<ArrayBuffer, string>;
+    exportGetBytes?: () => types.Result<Uint8Array, string>;
+    exportGetSerdeBytes?: () => types.Result<ArrayBuffer, string>;
     exportMultiplePrimitives?: (arg1: number, arg2: string) => bigint;
     exportPrimitiveBool?: (arg: boolean) => boolean;
     exportPrimitiveF32?: (arg: number) => number;
@@ -91,6 +93,7 @@ export type Exports = {
     exportFpUntaggedRaw?: (arg: Uint8Array) => Uint8Array;
     exportGenericsRaw?: (arg: Uint8Array) => Uint8Array;
     exportGetBytesRaw?: () => Uint8Array;
+    exportGetSerdeBytesRaw?: () => Uint8Array;
     exportMultiplePrimitivesRaw?: (arg1: number, arg2: Uint8Array) => bigint;
     exportPrimitiveBoolRaw?: (arg: boolean) => boolean;
     exportPrimitiveI16Raw?: (arg: number) => number;
@@ -253,6 +256,9 @@ export async function createRuntime(
             },
             __fp_gen_import_get_bytes: (): FatPtr => {
                 return serializeObject(importFunctions.importGetBytes());
+            },
+            __fp_gen_import_get_serde_bytes: (): FatPtr => {
+                return serializeObject(importFunctions.importGetSerdeBytes());
             },
             __fp_gen_import_multiple_primitives: (arg1: number, arg2_ptr: FatPtr): bigint => {
                 const arg2 = parseObject<string>(arg2_ptr);
@@ -443,6 +449,12 @@ export async function createRuntime(
         })(),
         exportGetBytes: (() => {
             const export_fn = instance.exports.__fp_gen_export_get_bytes as any;
+            if (!export_fn) return;
+
+            return () => parseObject<types.Result<Uint8Array, string>>(export_fn());
+        })(),
+        exportGetSerdeBytes: (() => {
+            const export_fn = instance.exports.__fp_gen_export_get_serde_bytes as any;
             if (!export_fn) return;
 
             return () => parseObject<types.Result<ArrayBuffer, string>>(export_fn());
@@ -658,6 +670,12 @@ export async function createRuntime(
         })(),
         exportGetBytesRaw: (() => {
             const export_fn = instance.exports.__fp_gen_export_get_bytes as any;
+            if (!export_fn) return;
+
+            return () => importFromMemory(export_fn());
+        })(),
+        exportGetSerdeBytesRaw: (() => {
+            const export_fn = instance.exports.__fp_gen_export_get_serde_bytes as any;
             if (!export_fn) return;
 
             return () => importFromMemory(export_fn());